exam_umarell/requirements.txt
flask==2.3.3
flask-restful
WTForms
requests
google-cloud-firestore
google-cloud-pubsub 
gunicorn
numpy